Added option --with-ide which used to generate ide file.
