+++ src/database/src/Commands/Ast/ModelUpdateVisitor.php
@@ -15,7 +15,6 @@ use Hyperf\Contract\Castable;
 use Hyperf\Contract\CastsAttributes;
 use Hyperf\Contract\CastsInboundAttributes;
 use Hyperf\Database\Commands\ModelOption;
-use Hyperf\Database\Model\Builder;
 use Hyperf\Database\Model\Collection;
 use Hyperf\Database\Model\Model;
 use Hyperf\Database\Model\Relations\BelongsTo;
@@ -238,17 +237,6 @@ class ModelUpdateVisitor extends NodeVisitorAbstract
                 continue;
             }
 
-            if (Str::startsWith($method->getName(), 'scope') && $method->getName() !== 'scopeQuery') {
-                $name = Str::camel(substr($method->getName(), 5));
-                if (! empty($name)) {
-                    $args = $method->getParameters();
-                    // Remove the first ($query) argument
-                    array_shift($args);
-                    $this->setMethod($name, [Builder::class, $method->getDeclaringClass()->getName()], $args);
-                }
-                continue;
-            }
-
             if ($method->getNumberOfParameters() > 0) {
                 continue;
             }
@@ -337,17 +325,6 @@ class ModelUpdateVisitor extends NodeVisitorAbstract
             $this->properties[$name]['write'] = $write;
         }
         $this->properties[$name]['priority'] = $priority;
-    }
-
-    protected function setMethod(string $name, array $type = [], array $arguments = [])
-    {
-        $methods = array_change_key_case($this->methods, CASE_LOWER);
-
-        if (! isset($methods[strtolower($name)])) {
-            $this->methods[$name] = [];
-            $this->methods[$name]['type'] = implode('|', $type);
-            $this->methods[$name]['arguments'] = $arguments;
-        }
     }
 
     protected function getProperty($column): array
